style: change modal style and rename
