--- src/admin/components/elements/BaseDialog/types.ts
@@ -1,12 +1,0 @@
-type Action = {
-  action: () => void;
-  label: string;
-};
-
-export type Props = {
-  open: boolean;
-  title: string;
-  body: string;
-  confirm: Action;
-  cancel: Action;
-};
