+++ src/admin/pages/Trash/RestoreButton/index.tsx
@@ -4,7 +4,7 @@ import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { logger } from '../../../../utilities/logger.js';
 import { IconButton } from '../../../@extended/components/IconButton/index.js';
-import { BaseDialog } from '../../../components/elements/BaseDialog/index.js';
+import { ModalDialog } from '../../../components/elements/ModalDialog/index.js';
 import { Icon } from '../../../components/elements/Icon/index.js';
 import { ComposeWrapper } from '../../../components/utilities/ComposeWrapper/index.js';
 import { TrashContextProvider, useTrash } from '../Context/index.js';
@@ -33,7 +33,7 @@ export const RestoreButtonImpl: React.FC<Props> = ({ postId, onRestored }) => {
 
   return (
     <>
-      <BaseDialog
+      <ModalDialog
         open={openRestore}
         title={t('dialog.confirm_post_restore_title')}
         body={t('dialog.confirm_post_restore')}
